requirements.txt
boto3
python-dotenv
socceraction
geopy
python-levenshtein
pandas
numpy
xgboost
scikit-learn
tqdm
unidecode
fuzzywuzzy